users can only read, update and delete their own tasks

+++ routes/api.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\ProjectTasksController;
 use App\Http\Controllers\TaskController;
 use App\Http\Controllers\UserTasksController;
 use App\Http\Middleware\AuthMiddleware;
+use App\Http\Middleware\EnsureTaskPermission;
 use Illuminate\Support\Facades\Route;
 
 Route::post('/login', [AuthController::class, 'login']);
@@ -15,10 +16,10 @@ Route::middleware(AuthMiddleware::class)
     ->group(function () {
         Route::get('/tasks', 'index');
         Route::get('/tasks/overdue', 'overdue');
-        Route::get('/tasks/{id}', 'show');
+        Route::get('/tasks/{id}', 'show')->middleware(EnsureTaskPermission::class);
         Route::post('/tasks', 'store');
-        Route::patch('/tasks/{id}', 'update');
-        Route::delete('/tasks/{id}', 'destroy');
+        Route::patch('/tasks/{id}', 'update')->middleware(EnsureTaskPermission::class);
+        Route::delete('/tasks/{id}', 'destroy')->middleware(EnsureTaskPermission::class);
 });
 
 Route::get('/users/{id}/tasks', UserTasksController::class)
